Add files via upload
first set of data pull from the website Seeks. One problem occurs is that the data from some columns do not seem to match each other. This might be because the different length of many lists that I created. I will try to fix this by appending data into a set of dictionary instead of storing them in many lists